+++ lib/onfido/null_logger.rb
@@ -1,5 +1,0 @@
-module Onfido
-  class NullLogger
-    def <<(*args); end
-  end
-end