+++ spec/integrations/address_spec.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 describe Onfido::Address do
-  subject(:api) { Onfido::API.new }
+  include_context 'fake onfido api'
 
   describe '#all' do
     it 'returns the addresses matching the postcode' do
-      response = api.address.all('SW1 4NG')
+      response = onfido.address.all('SW1 4NG')
 
       expect(response['addresses'].count).to eq(2)
     end