+++ spec/integrations/document_spec.rb
@@ -1,7 +1,11 @@
+# frozen_string_literal: true
+
 require 'tempfile'
 
 describe Onfido::Document do
-  subject(:document) { described_class.new }
+  include_context 'fake onfido api'
+
+  subject(:document) { onfido.document }
 
   describe '#create' do
     let(:params) do
@@ -32,8 +36,8 @@ describe Onfido::Document do
       let(:file) { 'https://onfido.com/images/logo.png' }
 
       it 'raises an ArgumentError' do
-        expect { document.create(**params) }.
-          to raise_error(ArgumentError, /must be a `File`-like object/)
+        expect { document.create(**params) }
+          .to raise_error(ArgumentError, /must be a `File`-like object/)
       end
     end
   end