+++ spec/integrations/extraction_spec.rb
@@ -1,7 +1,11 @@
+# frozen_string_literal: true
+
 require 'tempfile'
 
 describe Onfido::Extraction do
-  subject(:extraction) { described_class.new }
+  include_context 'fake onfido api'
+
+  subject(:extraction) { onfido.extraction }
 
   describe '#create' do
     let(:params) do
@@ -11,7 +15,7 @@ describe Onfido::Extraction do
     end
 
     it 'creates a new extraction' do
-      response = extraction.create(params)
+      response = extraction.create(**params)
 
       expect(response['document_id']).to eq('7568415-123123-123123')
     end